sail-intelligence/backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ── Async HTTP ────────────────────────────────────────────────────────────────
httpx==0.27.2
aiohttp==3.10.10

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# ── Validation / settings ─────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.6.1

# ── Task queue ────────────────────────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.2.0

# ── Vector DB ─────────────────────────────────────────────────────────────────
pinecone-client==5.0.1

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.2
cryptography==43.0.3

# ── LLM / Agents ──────────────────────────────────────────────────────────────
langgraph==0.2.48
langchain==0.3.7
langchain-openai==0.2.8
langchain-anthropic==0.3.0
openai==1.54.4
anthropic==0.39.0
groq==0.11.0

# ── Resilience ────────────────────────────────────────────────────────────────
tenacity==9.0.0
circuitbreaker==2.0.0

# ── Logging / observability ───────────────────────────────────────────────────
structlog==24.4.0

# ── Alerting ──────────────────────────────────────────────────────────────────
httpx==0.27.2   # reused — Telegram/WhatsApp webhook dispatch

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
orjson==3.10.11
ujson==5.10.0
